flask-service/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
